Initial configuration for devcontainers
This approach is completely different from how it was previously
implemented. Some key points to highlight:

0. From Docker to Devcontainers:

While Devcontainers are still a subset of Docker, a significant part of
the structure has been modified to adopt a devcontainer-first structure.
Although this might not necessarily be incompatible with Docker (as they
are an extension), supporting both approaches would be somewhat
cumbersome. Therefore, as an initial MVP, it has been temporarily
discarded.

1. Odoo: From binary to source code

Odoo is now executed using odoo-bin (source code) instead of the
installable binary. This allows direct control over Odoo's codebase
(which is useful for development) and enables the LSP to resolve
directly against it.

As an anecdote, I tried having a 'vendor' with the codebase to resolve
the LSP issue. However, since it turned out to be necessary anyway, I
decided it was better to execute Odoo directly from the source code and
avoid strange and impractical solutions.

2. Devcontainers prepared for VSCode (sorry, Neovim):

Despite exploring multiple options to use Devcontainers with Neovim, no
solution seemed good enough to justify not using VSCode (or any of its
forks).

Devcontainers is a technology developed by Microsoft, and while it is
now an open protocol that any IDE or similar tool can implement, it is
quite problematic in Neovim because Neovim lacks a Client-Server
architecture like VSCode. The most optimal option, in my opinion, is to
run Nvim directly inside the container. While this setup more or less
works, I found it too inconvenient due to the inability to use the
clipboard. Although there are alternatives, such as using the clipboard
of Zellij or Tmux or directly the system's clipboard depending on the
terminal/protocol supported, it does not feel quite right. It causes
non-trivial issues with formatting, registers, etc.

For the moment, at least while working on Dusarocker projects in Odoo,
the most sensible solution seems to be using VSCode or its forks.